requirements.txt
easyocr
imutils
opencv-python
matplotlib
paho-mqtt
flask